checkbox.rs: fix character type translations

+++ src/components/checkbox.rs
@@ -14,7 +14,7 @@ pub struct Checkbox {
 }
 
 impl Checkbox {
-    pub fn new(left: i32, top: i32, text: &str, def_value: i8, seq: &[u8])
+    pub fn new(left: i32, top: i32, text: &str, def_value: u8, seq: &[u8])
             -> Checkbox {
         #[link(name="newt")]
         extern "C" {
@@ -22,29 +22,32 @@ impl Checkbox {
                             defValue: c_char, seq: *const c_char,
                             result: *mut c_char) -> c_component;
         }
+
         let c_text = CString::new(text).unwrap();
         let s_seq = String::from_utf8_lossy(seq);
         let c_seq = CString::new(s_seq.into_owned()).unwrap();
         Checkbox {
             co: unsafe {
-                newtCheckbox(left, top, c_text.as_ptr(), def_value,
+                newtCheckbox(left, top, c_text.as_ptr(), def_value as i8,
                              c_seq.as_ptr(), ptr::null_mut())
             }
         }
     }
 
-    pub fn get_value(&self) -> i8 {
+    pub fn get_value(&self) -> u8 {
         #[link(name="newt")]
         extern "C" { fn newtCheckboxGetValue(co: c_component) -> c_char; }
-        unsafe { newtCheckboxGetValue(self.co) }
+
+        unsafe { newtCheckboxGetValue(self.co) as u8 }
     }
 
-    pub fn set_value(&self, value: i8) {
+    pub fn set_value(&self, value: u8) {
         #[link(name="newt")]
         extern "C" {
             fn newtCheckboxSetValue(co: c_component, value: c_char);
         }
-        unsafe { newtCheckboxSetValue(self.co, value); }
+
+        unsafe { newtCheckboxSetValue(self.co, value as i8); }
     }
 
     pub fn set_flags(&self, flags: i32, sense: FlagsSense) {
@@ -53,6 +56,7 @@ impl Checkbox {
             fn newtCheckboxSetFlags(co: c_component, flags: c_int,
                                     sense: FlagsSense);
         }
+
         unsafe { newtCheckboxSetFlags(self.co, flags, sense); }
     }
 }
